Don't share the last seen at session variables
We picked up a bug in development, where we noticed that it was possible
to extend an admin session by visiting a page on the public side of the site.

The seperates out the management of `last_seen_at` into two parts - one to
handle admins, and one to handle public users. When a user carries out an
action on the public side, this updates `session[:last_seen_at]` and when a
user carries out an action on the admin side, this updates
`session[:admin_last_seen_at]`.

I've also seperated out the `SessionsController`s, which are used when
a user using Javascript opts to extend their session. This means that both
the admin and application layouts now pass an additional
`refresh_session_path` variable, which is sent to the JS via data attributes,
and tells the frontend code which URL to send a request to.
