+++ app/assets/javascripts/components/timeout_dialog.js
@@ -8,6 +8,7 @@
   var continueButtonElement = document.getElementById("continue_session");
   var warningTimeInMinutes = dialogElement.getAttribute("data-warning-in-minutes");
   var timeoutInMinutes = dialogElement.getAttribute("data-timeout-in-minutes");
+  var refreshSessionPath = dialogElement.getAttribute("data-refresh-session-path");
   var timeoutInMilliseconds = (timeoutInMinutes - warningTimeInMinutes) * 60 * 1000;
   var timeoutPath = dialogElement.getAttribute("data-timeout-path");
   var dialog = new window.A11yDialog(dialogElement);
@@ -23,7 +24,7 @@
 
   function refreshSession() {
     Rails.ajax({
-      url: "<%= refresh_session_path %>",
+      url: refreshSessionPath,
       type: "get",
       success: function() {
         clearInterval(window.sessionTimeoutTimer);
