+++ app/controllers/admin/sessions_controller.rb
@@ -1,0 +1,9 @@
+module Admin
+  class SessionsController < BaseAdminController
+    def refresh
+      end_expired_admin_sessions
+
+      head :ok
+    end
+  end
+end
