+++ app/controllers/base_public_controller.rb
@@ -3,6 +3,7 @@ class BasePublicController < ApplicationController
 
   helper_method :current_policy_routing_name, :claim_timeout_in_minutes
   before_action :end_expired_claim_sessions
+  after_action :update_last_seen_at
 
   private
 
@@ -16,5 +17,9 @@ class BasePublicController < ApplicationController
       clear_claim_session
       redirect_to timeout_claim_path(policy_routing_name_for_redirect)
     end
+  end
+
+  def update_last_seen_at
+    session[:last_seen_at] = Time.zone.now
   end
 end
