+++ app/controllers/sessions_controller.rb
@@ -1,10 +1,8 @@
-class SessionsController < ApplicationController
-  include ClaimSessionTimeout
-  include AdminSessionTimeout
+class SessionsController < BasePublicController
+  skip_before_action :end_expired_claim_sessions
 
   def refresh
     clear_claim_session if claim_session_timed_out?
-    end_expired_admin_sessions
 
     head :ok
   end
