+++ app/views/application/_timeout_dialog.html.erb
@@ -6,6 +6,7 @@
       "timeout-in-minutes" => timeout_in_minutes,
       "warning-in-minutes" => timeout_warning_in_minutes,
       "timeout-path" => path_on_timeout,
+      "refresh-session-path" => refresh_session_path,
     },
   ) do %>
   <div class="dialog-overlay" tabindex="-1" data-a11y-dialog-hide=""></div>
